requirements.txt
streamlit>=1.28.0
openai>=1.0.0
tavily-python>=0.3.0
python-dotenv>=1.0.0
neo4j>=5.0.0
reka-api>=1.0.0
requests>=2.31.0
pandas>=2.0.0
plotly>=5.0.0
